+++ backend/app/core/webhooks/webhook.py
@@ -16,6 +16,7 @@ async def call_webhook(
     content: str, method: str, url: str, query: str, status_code: int
 ) -> None:
     webhook_url = settings.WEBHOOK_URL
+    logger.warning("Using deprecated webhook v1, please switch to v2")
     try:
         async with httpx.AsyncClient() as client:
             message = {
@@ -39,7 +40,11 @@ async def call_webhook(
 
             resp = await client.post(webhook_url, json=message, headers=headers)
             if resp.status_code != 200:
-                logger.warning(f"Failed to send notification to {webhook_url}")
+                logger.warning(f"Response received {webhook_url}")
+                logger.warning(
+                    f"Failed to send notification to {webhook_url} \n response: {resp.text} \n Message sent: {message} \n Headers: {headers}"
+                )
+
     except Exception as e:
         logger.warning(f"Failed to send notification to {webhook_url}", e)